Change element type of time to LONG

+++ src/main/java/de/uni_koblenz/west/koral/master/statisticsDB/impl/multi_file/log/StorageLogWriter.java
@@ -43,7 +43,7 @@ public class StorageLogWriter {
 		accessEventLayout.put(KEY_ACCESS_FILESIZE, ElementType.INTEGER);
 		accessEventLayout.put(KEY_ACCESS_CACHEHIT, ElementType.BIT);
 		accessEventLayout.put(KEY_ACCESS_FOUND, ElementType.BIT);
-		accessEventLayout.put(KEY_ACCESS_TIME, ElementType.INTEGER);
+		accessEventLayout.put(KEY_ACCESS_TIME, ElementType.LONG);
 		rowLayouts.put(StorageLogEvent.READWRITE.ordinal(), accessEventLayout);
 
 		Map<String, ElementType> blockFlushEventLayout = new TreeMap<>();
